Fix bug in jenkins cleanup code.

+++ jenkins.sh
@@ -508,10 +508,7 @@ jenkins_delete_files()
         allFiles+=("/tmp/${file}")
     done
 
-    # If there's nothing to do just return success immediately
-    [[ ${#allFiles[@]} -eq 0 ]] && return 0
-    
-    ssh_jenkins "rm -f ${allFiles[@]}"
+    [[ ${#allFiles[@]} -gt 0 ]] && ssh_jenkins "rm -f ${allFiles[@]}" || return 0
 }
 
 return 0
